Fix ac-expand problem at the end of buffer.
When at eob, ac menu is distorted.  The cursor and inline overlay are moved
beyond the ac-menu. Consequently ac-expand (tab) doesn't work when it is invoked
repeatedly. This happens because the point is moved out of the ac-prefix-overlay
region on first [tab]. This is extremely harmful in the shell and comint buffers
when editing is always at eob.

There is already a workaround in auto-complete.el for the case when ac-point
== (point-min) == (point-max). This workaround should be called always when
point is at eob.