Fix blinking cursor when no match is found
ac-cursor-color was set by ac-start() even though it is only relevant when fuzzy matching is on, and additionally only when ac-fuzzy-cursor-color is non-nil.  Because it was getting set,  ac-cleanup() was always calling set-cursor-color().  That ought to be a no-op when fuzzy matching is off (since the cursor color was never actually changed) but at least on OSX this causes a brief but noticeable cursor blink (in all buffers).

I believe this should fix:
- Issue #351
- http://stackoverflow.com/questions/21793920/stop-emacs-auto-complete-from-flashing-cursor-in-other-windows
- http://stackoverflow.com/questions/24227855/auto-complete-mode-on-emacs-for-mac-os-x

To test, I opened emacs on OSX and typed aaaaaaaaaaaaa in the scratch buffer.  Previously, the cursor would blink with every keystroke.  With this patch it doesn't.